rag_local/requirements.txt
python-dotenv
requests
sentence-transformers
faiss-cpu
numpy
responses
pytest
